Default auto (#42805)
* default to `"auto"` dtype

* the actual change?

* up?

* style

* up?

* only sam models were broken with this

* fix sams

* update

* fix sam2 now

* up

* this?

* proper fix

* lol

* fix

* fixes

* nit

* fix

* fix copies

* fixes

* fix bigbird

* revert one bit

+++ src/transformers/models/edgetam/modeling_edgetam.py
@@ -393,7 +393,7 @@ class EdgeTamVisionNeck(nn.Module):
         n = len(self.convs) - 1
         for i in range(n, -1, -1):
             lateral_features = hidden_states[i].permute(0, 3, 1, 2)
-            lateral_features = self.convs[n - i](lateral_features)
+            lateral_features = self.convs[n - i](lateral_features.to(self.convs[i].weight.dtype))
             if i not in self.fpn_top_down_levels or i == n:
                 prev_features = lateral_features
             else:
